# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API for AuthService",
        "title": "AuthService API",
        "contact": {
            "name": "Andrei Venski",
            "url": "https://github.com/andrew967",
            "email": "venskiandrei32@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1/auth",
    "paths": {
        "/tokens": {
            "post": {
                "description": "Get new tokens : access and refresh",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get New Tokens",
                "operationId": "getNewTokens",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's IP Address",
                        "name": "X-Forwarded-For",
                        "in": "header"
                    },
                    {
                        "description": "User Info",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GetNewTokenData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tokens/refresh": {
            "post": {
                "description": "Refresh access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Refresh Tokens",
                "operationId": "refreshToken",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's IP Address",
                        "name": "X-Forwarded-For",
                        "in": "header"
                    },
                    {
                        "description": "Refresh Token Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RefreshData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "models.GetNewTokenData": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.RefreshData": {
            "type": "object",
            "required": [
                "refresh_token",
                "refresh_token_id"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                },
                "refresh_token_id": {
                    "type": "string"
                }
            }
        },
        "models.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token_id": {
                    "type": "string"
                }
            }
        }
    }
}